requirements.txt
 
# Core Dependencies
python-telegram-bot[job-queue]==20.7
tvdatafeed==2.1.0
yfinance==0.2.33

# Data Analysis & Technical Indicators
pandas==2.1.4
numpy==1.24.3
pandas-ta==0.3.14b0
scikit-learn==1.3.2
scipy==1.11.4

# Charting & Visualization
mplfinance==0.12.10b0
matplotlib==3.8.2
Pillow==10.1.0

# Task Scheduling
APScheduler==3.10.4

# Configuration & Utils
python-dotenv==1.0.0
requests==2.31.0

# Optional: TA-Lib for advanced technical analysis
# ta-lib==0.4.28  # Uncomment if you want additional TA indicators

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1